cpufreq: stats: Handle the case when trans_table goes beyond PAGE_SIZE
On platforms with large number of Pstates, the transition table, which
is a NxN matrix, can overflow beyond the PAGE_SIZE boundary.

This can be seen on POWER9 which has 100+ Pstates.

As a result, each time the trans_table is read for any of the CPUs, we
will get the following error.

fill_read_buffer: show+0x0/0xa0 returned bad count

This patch ensures that in case of an overflow, we print a warning
once in the dmesg and return FILE TOO LARGE error for this and all
subsequent accesses of trans_table.

Change-Id: I4d22a81fcad037e887b4ddf2999904b321401441
Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Git-repo: https://android.googlesource.com/kernel/common/
Git-commit: f7bc9b209e27c0b617378400136cc663a6314d0c
Signed-off-by: Rahul Shahare <rshaha@codeaurora.org>