[BACKPORT]fs: only set S_VERSION when updating times if necessary
We only really need to update i_version if someone has queried for it
since we last incremented it. By doing that, we can avoid having to
update the inode if the times haven't changed.

If the times have changed, then we go ahead and forcibly increment the
counter, under the assumption that we'll be going to the storage
anyway, and the increment itself is relatively cheap.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Jan Kara <jack@suse.cz>

@THEBOSS619:- use "MS_LAZYTIME" instead of "SB_LAZYTIME" and Adapted to FIVE samsung security changes [reason I havn't removed samsung changes is because to prevent regression on write performance]