mm/hpa: fix dependency to lowmemorykiller
lowmemorykiller is removed from the tree by the following commit:
915e70f ("staging, android: remove lowmemory killer from the tree")

Instead HPA killer borrows victim selection logic from oom_killer.
oom_killer selectes a victim task with the combination of the total
rss of the task and oom_adj_score. But HPA killer selects the task
with the largest rss among the tasks with the same oom_adj_score that
is also the largest to select a task with less panelty.

Change-Id: I0ef31e5c44faad9c6216dbdbd085082167b3f256
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>