block: wait for a queue to unfreeze, or unset the PREEMPT_ONLY flag
@THEBOSS619:- by reverting this commit it should make sure that during
queuing it doesn't get stuck and makes sure that the tasks continue it's
operations normally without issues.

either way this commit is targeted for upstreamed block driver which isn't
the case on this current kernel state... will look it up on the future
once we reached to 4.14k+ upstreamed block drivers commits.. although I
doubt it.

This reverts commit 3118ceb456200d73179ddf29f2b95f59e35465f9.